+++ examples/scrollXY.tsx
@@ -29,17 +29,30 @@ const data = [
   { a: '133', c: 'edd12221 edd12221', d: 2, key: '9' },
 ];
 
-const Demo = () => (
-  <div>
-    <h2>Scroll X/Y</h2>
-    <Table
-      useFixedHeader
-      style={{ width: 800 }}
-      scroll={{ x: 1500, y: 300 }}
-      columns={columns}
-      data={data}
-    />
-  </div>
-);
+const Demo = () => {
+  const [visible, setVisible] = React.useState(true);
+  return (
+    <div>
+      <h2>Scroll X/Y</h2>
+      <button
+        type="button"
+        onClick={() => {
+          setVisible(!visible);
+        }}
+      >
+        Trigger Visible
+      </button>
+      <div style={{ display: visible ? undefined : 'none' }}>
+        <Table
+          useFixedHeader
+          style={{ width: 800 }}
+          scroll={{ x: 1500, y: 300 }}
+          columns={columns}
+          data={data}
+        />
+      </div>
+    </div>
+  );
+};
 
 export default Demo;