+++ src/interface.ts
@@ -91,7 +91,8 @@ export interface ColumnType<RecordType> extends ColumnSharedType<RecordType> {
 
 export type ColumnsType<RecordType = unknown> = (
   | ColumnGroupType<RecordType>
-  | ColumnType<RecordType>)[];
+  | ColumnType<RecordType>
+)[];
 
 export type GetRowKey<RecordType> = (record: RecordType, index?: number) => Key;
 
@@ -212,4 +213,5 @@ export type TriggerEventHandler<RecordType> = (
 export interface TableSticky {
   offsetHeader?: number;
   offsetScroll?: number;
+  getContainer?: () => Window | HTMLElement;
 }