+++ tests/FixedHeader.spec.js
@@ -1,10 +1,28 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { act } from 'react-dom/test-utils';
+import { spyElementPrototype } from 'rc-util/lib/test/domHook';
 import Table, { INTERNAL_COL_DEFINE } from '../src';
 
 describe('Table.FixedHeader', () => {
-  it('switch column', () => {
+  let domSpy;
+  let visible = true;
+
+  beforeAll(() => {
+    domSpy = spyElementPrototype(HTMLElement, 'offsetParent', {
+      get: () => (visible ? {} : null),
+    });
+  });
+
+  beforeEach(() => {
+    visible = true;
+  });
+
+  afterAll(() => {
+    domSpy.mockRestore();
+  });
+
+  it('should work', () => {
     jest.useFakeTimers();
     const col1 = { dataIndex: 'light', width: 100 };
     const col2 = { dataIndex: 'bamboo', width: 200 };
@@ -17,21 +35,9 @@ describe('Table.FixedHeader', () => {
       />,
     );
 
-    wrapper
-      .find('ResizeObserver')
-      .at(0)
-      .props()
-      .onResize({ width: 100, offsetWidth: 100 });
-    wrapper
-      .find('ResizeObserver')
-      .at(1)
-      .props()
-      .onResize({ width: 200, offsetWidth: 200 });
-    wrapper
-      .find('ResizeObserver')
-      .at(2)
-      .props()
-      .onResize({ width: 0, offsetWidth: 0 });
+    wrapper.find('ResizeObserver').at(0).props().onResize({ width: 100, offsetWidth: 100 });
+    wrapper.find('ResizeObserver').at(1).props().onResize({ width: 200, offsetWidth: 200 });
+    wrapper.find('ResizeObserver').at(2).props().onResize({ width: 0, offsetWidth: 0 });
 
     act(() => {
       jest.runAllTimers();
@@ -41,41 +47,16 @@ describe('Table.FixedHeader', () => {
     expect(wrapper.find('.rc-table-header table').props().style.visibility).toBeFalsy();
 
     expect();
-    expect(
-      wrapper
-        .find('colgroup col')
-        .at(0)
-        .props().style.width,
-    ).toEqual(100);
-    expect(
-      wrapper
-        .find('colgroup col')
-        .at(1)
-        .props().style.width,
-    ).toEqual(200);
-    expect(
-      wrapper
-        .find('colgroup col')
-        .at(2)
-        .props().style.width,
-    ).toEqual(0);
+    expect(wrapper.find('colgroup col').at(0).props().style.width).toEqual(100);
+    expect(wrapper.find('colgroup col').at(1).props().style.width).toEqual(200);
+    expect(wrapper.find('colgroup col').at(2).props().style.width).toEqual(0);
 
     // Update columns
     wrapper.setProps({ columns: [col2, col1] });
     wrapper.update();
 
-    expect(
-      wrapper
-        .find('colgroup col')
-        .at(0)
-        .props().style.width,
-    ).toEqual(200);
-    expect(
-      wrapper
-        .find('colgroup col')
-        .at(1)
-        .props().style.width,
-    ).toEqual(100);
+    expect(wrapper.find('colgroup col').at(0).props().style.width).toEqual(200);
+    expect(wrapper.find('colgroup col').at(1).props().style.width).toEqual(100);
 
     jest.useRealTimers();
   });
@@ -95,22 +76,12 @@ describe('Table.FixedHeader', () => {
       />,
     );
 
-    expect(
-      wrapper
-        .find('table')
-        .last()
-        .find('colgroup col')
-        .first()
-        .props().className,
-    ).toEqual('test-internal');
-    expect(
-      wrapper
-        .find('table')
-        .first()
-        .find('colgroup col')
-        .first()
-        .props().className,
-    ).toEqual('test-internal');
+    expect(wrapper.find('table').last().find('colgroup col').first().props().className).toEqual(
+      'test-internal',
+    );
+    expect(wrapper.find('table').first().find('colgroup col').first().props().className).toEqual(
+      'test-internal',
+    );
   });
 
   it('show header when data is null', () => {
@@ -141,5 +112,63 @@ describe('Table.FixedHeader', () => {
     expect(wrapper.find('.rc-table-header table').props().style).toEqual(
       expect.objectContaining({ visibility: null }),
     );
+  });
+
+  it('rtl', () => {
+    const wrapper = mount(
+      <Table
+        columns={[{ dataIndex: 'light', width: 100 }]}
+        data={[{ key: 0, light: 'bamboo' }]}
+        direction="rtl"
+        scroll={{
+          y: 100,
+        }}
+      />,
+    );
+
+    expect(wrapper.find('Header').props().stickyOffsets).toEqual(
+      expect.objectContaining({
+        isSticky: false,
+        left: [expect.anything(), expect.anything()],
+      }),
+    );
+  });
+
+  it('invisible should not change width', () => {
+    jest.useFakeTimers();
+
+    const col1 = { dataIndex: 'light', width: 93 };
+    const wrapper = mount(
+      <Table
+        columns={[col1]}
+        data={[{ light: 'bamboo', bamboo: 'light', key: 1 }]}
+        scroll={{ y: 10 }}
+      />,
+    );
+
+    wrapper.find('ResizeObserver').at(0).props().onResize({ width: 93, offsetWidth: 93 });
+    act(() => {
+      jest.runAllTimers();
+      wrapper.update();
+    });
+
+    expect(wrapper.find('FixedHeader col').first().props().style).toEqual(
+      expect.objectContaining({ width: 93 }),
+    );
+
+    // Hide Table should not modify column width
+    visible = false;
+
+    wrapper.find('ResizeObserver').at(0).props().onResize({ width: 0, offsetWidth: 0 });
+    act(() => {
+      jest.runAllTimers();
+      wrapper.update();
+    });
+
+    expect(wrapper.find('FixedHeader col').first().props().style).toEqual(
+      expect.objectContaining({ width: 93 }),
+    );
+
+    jest.useRealTimers();
   });
 });