+++ tests/Scroll.spec.js
@@ -88,18 +88,17 @@ describe('Table.Scroll', () => {
     jest.runAllTimers();
     // Use `onScroll` directly since simulate not support `currentTarget`
     act(() => {
-      wrapper
-        .find('.rc-table-header')
-        .props()
-        .onScroll({
-          currentTarget: {
-            scrollLeft: 10,
-            scrollWidth: 200,
-            clientWidth: 100,
-          },
-        });
+      const headerDiv = wrapper.find('div.rc-table-header').instance();
+
+      const wheelEvent = new WheelEvent('wheel');
+      Object.defineProperty(wheelEvent, 'deltaX', {
+        get: () => 10,
+      });
+
+      headerDiv.dispatchEvent(wheelEvent);
+      jest.runAllTimers();
     });
-    jest.runAllTimers();
+
     expect(setScrollLeft).toHaveBeenCalledWith(undefined, 10);
     setScrollLeft.mockReset();